Keine willkürliche Farbgrenze mehr: bis 375, so viele wie der Katalog hat
Die Farbanzahl war bei 48 gedeckelt. Der Katalog hat 375 Garne, und mehr
als das zu verlangen ergibt keinen Sinn – weniger zu erlauben aber auch
nicht: ob ein Muster mit 300 Farben zu sticken ist, entscheidet die
Nutzerin und nicht das Programm.

Damit das trägt, mussten drei Dinge umgebaut werden:

Ein Feld hält zwei Byte statt einem. Mit einem Byte war bei 255 Farben
Schluss, und die 255 war schon für „wird nicht gestickt" vergeben. Jetzt
laufen die Palettenindizes von 0 bis 1022, LEER ist die 1023. Die beiden
gespeicherten Formate (Raster-RLE und Motive) tragen dafür eine
Fassungsnummer und lesen die vorige weiter: wer ein Muster gespeichert
hat, findet es unverändert wieder.

Aus der Abstandstabelle ist eine Abstandsliste geworden. Vorher stand für
jedes Feld der Abstand zu jeder Palettenfarbe in einer Tabelle: bei
160.000 Feldern und 48 Farben 30 MB, bei 375 Farben aber 240 MB – das
überlebt kein Tablet. Jetzt stehen dort nur noch die zwölf
nächstliegenden Farben je Feld, rund 12 MB, unabhängig von der
Farbanzahl. Das Ergebnis bleibt dasselbe: für jede Farbe, die in der
Nachbarschaft eines Feldes nicht vorkommt, ist die Strafe der Glättung
gleich hoch, also kann unter ihnen nur die farbtreueste gewinnen – und
die steht in der Liste. Gegengerechnet gegen die alte Fassung auf Bildern
in voller Größe: Feld für Feld dasselbe Muster, dieselben Kennzahlen, bei
20 und 48 Farben auch dieselbe oder eine kürzere Rechenzeit.

Das k-Means bekommt ein Durchlaufbudget. Ein Durchlauf kostet Felder ×
Farben Abstände; bei 375 Farben in voller Größe wären vierzig Durchläufe
über acht Sekunden. Zwischen zwölf und vierzig Durchläufen liegen dort
gemessene 0,002 ΔE – unsichtbar. Bis 48 Farben bleibt alles wie bisher.

Ein Muster in voller Größe mit 375 Farben braucht damit rund fünf
Sekunden (48 Farben: knapp vier).

In der Oberfläche wächst die Schrittweite mit der Zahl: von 2 auf 375
sind es 29 Tipps auf „Mehr" statt 187. „Weniger" führt immer wieder auf
denselben Wert zurück.

+++ src/lib/druck/pdf.ts
@@ -59,7 +59,7 @@ export type Druckauftrag = {
   name: string;
   breite: number;
   hoehe: number;
-  raster: Uint8Array;
+  raster: Uint16Array;
   /**
    * Nur die Farben, die im Muster wirklich vorkommen. Angesprochen wird sie
    * über `index` und nie über die Stelle in der Liste – Felder, die nicht
